refactor: organize tests by service and then by control vs data
In preparation for introducing per-endpoint make targets, we organize
the integration tests by service first and then control vs data
after. We add fully qualified namespaces for this organization so we
can filter on the command line.

We also add namespaces for `Integration` vs `Unit`.
