+++ tests/Integration/Momento.Sdk.Tests/Cache/Data/CacheEagerConnectionTest.cs
@@ -3,7 +3,7 @@ using Momento.Sdk.Auth;
 using Momento.Sdk.Config;
 using Momento.Sdk.Config.Transport;
 
-namespace Momento.Sdk.Tests.Integration;
+namespace Momento.Sdk.Tests.Integration.Cache.Data;
 
 public class CacheEagerConnectionTest
 {
@@ -38,7 +38,7 @@ public class CacheEagerConnectionTest
         config = config.WithTransportStrategy(config.TransportStrategy
                                                     .WithGrpcConfig(grpcConfiguration)
                                                     .WithMaxConcurrentRequests(2));
-       
+
         // just validating that we can construct the client wh
         var client = new CacheClient(config, authProvider, defaultTtl);
         // still 2; clients shouldn't know we are doing 2/10 magic internally
@@ -52,7 +52,7 @@ public class CacheEagerConnectionTest
         var config = Configurations.Laptop.Latest(loggerFactory);
         var authProviderWithBadCacheEndpoint = authProvider.WithCacheEndpoint("cache.cell-external-beta-1.prod.a.momentohq.com:65000");
         Console.WriteLine($"Hello developer!  We are about to run a test that verifies that the cache client is still operational even if our eager connection (ping) fails.  So you will see the test log a warning message about that.  It's expected, don't worry!");
-        
+
         await Assert.ThrowsAsync<ConnectionException>(async () => await CacheClient.CreateAsync(config, authProviderWithBadCacheEndpoint, defaultTtl, TimeSpan.FromSeconds(2)));
     }
 }
