+++ tests/Integration/Momento.Sdk.Tests/Cache/Data/DictionaryTest.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Momento.Sdk.Internal.ExtensionMethods;
 using Momento.Sdk.Requests;
 
-namespace Momento.Sdk.Tests;
+namespace Momento.Sdk.Tests.Integration.Cache.Data;
 
 [Collection("CacheClient")]
 public class DictionaryTest : TestBase
