+++ tests/Integration/Momento.Sdk.Tests/Usings.cs
@@ -2,5 +2,6 @@ global using System;
 global using Momento.Sdk;
 global using Momento.Sdk.Exceptions;
 global using Momento.Sdk.Responses;
+global using Momento.Sdk.Tests.Integration;
 global using Xunit;
 global using Utils = Momento.Sdk.Tests.Integration.Utils;
